Begin combining World and Building classes
