+++ World/Building.cs
@@ -14,6 +14,10 @@ public class Building : Node
 	public static Dictionary<int, List<Tuple<int,int>>> RemoteLoadedChunks = new Dictionary<int, List<Tuple<int,int>>>();
 	public static GridClass Grid = new GridClass();
 
+	public static List<DroppedItem> DroppedItems = new List<DroppedItem>();
+
+	private static PackedScene DroppedItemScene;
+
 	public static Building Self;
 
 	Building()
@@ -21,6 +25,8 @@ public class Building : Node
 		if(Engine.EditorHint) {return;}
 
 		Self = this;
+
+		DroppedItemScene = GD.Load<PackedScene>("res://Items/DroppedItem.tscn");
 
 		Directory StructureDir = new Directory();
 		StructureDir.Open("res://World/Scenes/");
@@ -296,5 +302,38 @@ public class Building : Node
 			Branch.QueueFree();
 		}
 
+	}
+
+
+	//Should be able to be called without rpc yet only run on server
+	[Remote]
+	public void DropItem(Items.TYPE Type, Vector3 Position, Vector3 BaseMomentum)
+	{
+		if(Self.GetTree().GetNetworkPeer() != null)
+		{
+			if(Self.GetTree().IsNetworkServer())
+			{
+				Net.SteelRpc(Self, nameof(NetDropItem), Type, Position, BaseMomentum);
+				NetDropItem(Type, Position, BaseMomentum);
+			}
+			else
+			{
+				Self.RpcId(Net.ServerId, nameof(DropItem), Type, Position, BaseMomentum);
+			}
+		}
+	}
+
+
+	[Remote]
+	public void NetDropItem(Items.TYPE Type, Vector3 Position, Vector3 BaseMomentum)
+	{
+		DroppedItem ToDrop = DroppedItemScene.Instance() as DroppedItem;
+		ToDrop.Translation = Position;
+		ToDrop.Momentum = BaseMomentum;
+		ToDrop.Type = Type;
+		ToDrop.GetNode<MeshInstance>("MeshInstance").Mesh = Items.Meshes[Type];
+
+		DroppedItems.Add(ToDrop);
+		Game.StructureRoot.AddChild(ToDrop);
 	}
 }
