+++ project.godot
@@ -26,7 +26,6 @@ Console="*res://Scripting/Console.cs"
 Scripting="*res://Scripting/Scripting.cs"
 Net="*res://Net/Net.cs"
 Bindings="*res://Scripting/Bindings.cs"
-WorldAutoLoad="*res://World/World.cs"
 Items="*res://Items/Items.cs"
 Building="*res://World/Building.cs"
 Menu="*res://UI/Menu/Menu.cs"
